Mark DateFilter as public
Also fix typo in file name and move a function to a separate file.

JIRA: RAIL-3047
