+++ libs/sdk-ui-filters/src/DateFilter/DateFilterCore.tsx
@@ -9,7 +9,7 @@ import { IntlWrapper } from "@gooddata/sdk-ui";
 import { MediaQueries } from "../constants";
 import { DateFilterButtonLocalized } from "./DateFilterButtonLocalized/DateFilterButtonLocalized";
 import { DateFilterBody } from "./DateFilterBody/DateFilterBody";
-import { applyExcludeCurrentPeriod } from "./utils/PeriodExlusion";
+import { applyExcludeCurrentPeriod } from "./utils/PeriodExclusion";
 import { formatAbsoluteDate } from "./utils/Translations/DateFilterTitle";
 import { DEFAULT_DATE_FORMAT } from "./constants/Platform";
 import { filterVisibleDateFilterOptions, sanitizePresetIntervals } from "./utils/OptionUtils";
