+++ libs/sdk-ui-filters/src/DateFilter/tests/DateFilter.test.tsx
@@ -1,5 +1,4 @@
 // (C) 2019 GoodData Corporation
-import { testAPI } from "../DateFilter";
 import {
     createDateFilter,
     clickDateFilterButton,
@@ -42,8 +41,6 @@ import {
     clickRelativeFormFilter,
     isRelativeFormSelectMenuVisible,
 } from "./extendedDateFilters_helpers";
-import { IRelativeDateFilterPreset } from "@gooddata/sdk-backend-spi";
-import { IUiRelativeDateFilterForm } from "../interfaces";
 import { DEFAULT_DATE_FORMAT } from "../constants/Platform";
 import { verifyDateFormat } from "../DateFilterCore";
 
@@ -651,54 +648,5 @@ describe("DateFilter", () => {
 
             expect(getRelativeFormInputFromValue(wrapper)).toEqual("");
         });
-    });
-});
-
-describe("normalizeSelectedFilterOption", () => {
-    const sampleFloatingRangeFormOption: IUiRelativeDateFilterForm = {
-        from: -2,
-        to: 0,
-        granularity: "GDC.time.date",
-        localIdentifier: "RELATIVE_FORM",
-        type: "relativeForm",
-        name: "",
-        visible: true,
-    };
-
-    const sampleUnsortedFloatingRangeFormOption: IUiRelativeDateFilterForm = {
-        from: 99,
-        to: -2,
-        granularity: "GDC.time.date",
-        localIdentifier: "RELATIVE_FORM",
-        type: "relativeForm",
-        name: "",
-        visible: true,
-    };
-
-    const sampleFloatingRangePresetOption: IRelativeDateFilterPreset = {
-        from: -11,
-        to: 0,
-        granularity: "GDC.time.month",
-        localIdentifier: "LAST_12_MONTHS",
-        type: "relativePreset",
-        name: "",
-        visible: true,
-    };
-
-    it("should sort floating range from and to fields", () => {
-        const normalizedOption: IUiRelativeDateFilterForm = testAPI.normalizeSelectedFilterOption(
-            sampleUnsortedFloatingRangeFormOption,
-        ) as any;
-        expect(normalizedOption.from < normalizedOption.to).toBe(true);
-    });
-
-    it("should not update floating range from and to fields if already in correct order", () => {
-        const actual = testAPI.normalizeSelectedFilterOption(sampleFloatingRangeFormOption);
-        expect(actual).toBe(sampleFloatingRangeFormOption);
-    });
-
-    it("should leave other options as is", () => {
-        const actual = testAPI.normalizeSelectedFilterOption(sampleFloatingRangePresetOption);
-        expect(actual).toBe(sampleFloatingRangePresetOption);
     });
 });
