+++ libs/sdk-ui-filters/src/DateFilter/utils/FilterOptionNormalization.ts
@@ -1,0 +1,16 @@
+// (C) 2007-2021 GoodData Corporation
+import { DateFilterOption, isUiRelativeDateFilterForm } from "../interfaces";
+
+export function normalizeSelectedFilterOption(selectedFilterOption: DateFilterOption): DateFilterOption {
+    if (
+        isUiRelativeDateFilterForm(selectedFilterOption) &&
+        selectedFilterOption.from > selectedFilterOption.to
+    ) {
+        return {
+            ...selectedFilterOption,
+            from: selectedFilterOption.to,
+            to: selectedFilterOption.from,
+        };
+    }
+    return selectedFilterOption;
+}
