+++ libs/sdk-ui-filters/src/DateFilter/utils/tests/FilterOptionNormalization.tets.ts
@@ -1,0 +1,53 @@
+// (C) 2021 GoodData Corporation
+import { IRelativeDateFilterPreset } from "@gooddata/sdk-backend-spi";
+import { IUiRelativeDateFilterForm } from "../../interfaces";
+import { normalizeSelectedFilterOption } from "../FilterOptionNormalization";
+
+describe("normalizeSelectedFilterOption", () => {
+    const sampleFloatingRangeFormOption: IUiRelativeDateFilterForm = {
+        from: -2,
+        to: 0,
+        granularity: "GDC.time.date",
+        localIdentifier: "RELATIVE_FORM",
+        type: "relativeForm",
+        name: "",
+        visible: true,
+    };
+
+    const sampleUnsortedFloatingRangeFormOption: IUiRelativeDateFilterForm = {
+        from: 99,
+        to: -2,
+        granularity: "GDC.time.date",
+        localIdentifier: "RELATIVE_FORM",
+        type: "relativeForm",
+        name: "",
+        visible: true,
+    };
+
+    const sampleFloatingRangePresetOption: IRelativeDateFilterPreset = {
+        from: -11,
+        to: 0,
+        granularity: "GDC.time.month",
+        localIdentifier: "LAST_12_MONTHS",
+        type: "relativePreset",
+        name: "",
+        visible: true,
+    };
+
+    it("should sort floating range from and to fields", () => {
+        const normalizedOption: IUiRelativeDateFilterForm = normalizeSelectedFilterOption(
+            sampleUnsortedFloatingRangeFormOption,
+        ) as any;
+        expect(normalizedOption.from < normalizedOption.to).toBe(true);
+    });
+
+    it("should not update floating range from and to fields if already in correct order", () => {
+        const actual = normalizeSelectedFilterOption(sampleFloatingRangeFormOption);
+        expect(actual).toBe(sampleFloatingRangeFormOption);
+    });
+
+    it("should leave other options as is", () => {
+        const actual = normalizeSelectedFilterOption(sampleFloatingRangePresetOption);
+        expect(actual).toBe(sampleFloatingRangePresetOption);
+    });
+});
